using graph.describe
